LPS-198878 Move LockedLayoutorder and LockedLayoutType as inner enums of LockedLayoutsDisplayContext
