+++ modules/apps/layout/layout-locked-layouts-web/src/main/java/com/liferay/layout/locked/layouts/web/internal/display/context/LockedLayoutsSearchContainerManagementToolbarDisplayContext.java
@@ -12,8 +12,6 @@ import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemList;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItemListBuilder;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItemListBuilder;
-import com.liferay.layout.constants.LockedLayoutType;
-import com.liferay.layout.model.LockedLayoutOrder;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.language.LanguageUtil;
@@ -67,7 +65,8 @@ public class LockedLayoutsSearchContainerManagementToolbarDisplayContext
 			StringPool.BLANK
 		).setParameter(
 			"orderByCol",
-			LockedLayoutOrder.LockedLayoutOrderType.LAST_AUTOSAVE.getValue()
+			LockedLayoutsDisplayContext.LockedLayoutOrder.LockedLayoutOrderType.
+				LAST_AUTOSAVE.getValue()
 		).setParameter(
 			"orderByType", "desc"
 		).setParameter(
@@ -102,7 +101,7 @@ public class LockedLayoutsSearchContainerManagementToolbarDisplayContext
 
 	@Override
 	public List<LabelItem> getFilterLabelItems() {
-		LockedLayoutType lockedLayoutType =
+		LockedLayoutsDisplayContext.LockedLayoutType lockedLayoutType =
 			_lockedLayoutsDisplayContext.getLockedLayoutType();
 
 		if (lockedLayoutType == null) {
@@ -168,7 +167,9 @@ public class LockedLayoutsSearchContainerManagementToolbarDisplayContext
 			}
 		).build();
 
-		for (LockedLayoutType lockedLayoutType : LockedLayoutType.values()) {
+		for (LockedLayoutsDisplayContext.LockedLayoutType lockedLayoutType :
+				LockedLayoutsDisplayContext.LockedLayoutType.values()) {
+
 			dropdownItems.add(
 				DropdownItemBuilder.setActive(
 					Objects.equals(
@@ -186,14 +187,15 @@ public class LockedLayoutsSearchContainerManagementToolbarDisplayContext
 	}
 
 	private List<DropdownItem> _getOrderDropdownItems() {
-		LockedLayoutOrder lockedLayoutOrder =
+		LockedLayoutsDisplayContext.LockedLayoutOrder lockedLayoutOrder =
 			_lockedLayoutsDisplayContext.getLockedLayoutOrder();
 
 		return new DropdownItemList() {
 			{
-				for (LockedLayoutOrder.LockedLayoutOrderType
-						lockedLayoutOrderType :
-							LockedLayoutOrder.LockedLayoutOrderType.values()) {
+				for (LockedLayoutsDisplayContext.LockedLayoutOrder.
+						LockedLayoutOrderType lockedLayoutOrderType :
+							LockedLayoutsDisplayContext.LockedLayoutOrder.
+								LockedLayoutOrderType.values()) {
 
 					add(
 						dropdownItem -> {
